fix(readerSlug): moving readerSlug to the int mask package
