+++ packages/int-mask/package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "lodash": "4.17.21",
+    "md5": "2.3.0",
     "tslib": "2.7.0",
     "underscore.string": "3.3.6"
   },
@@ -25,6 +26,7 @@
     "@jest/globals": "29.7.0",
     "@pocket-tools/eslint-config": "workspace:*",
     "@types/jest": "29.5.12",
+    "@types/md5": "2.3.5",
     "@types/node": "^22.5.2",
     "jest": "29.7.0",
     "ts-jest": "29.2.5",
