+++ packages/int-mask/src/index.ts
@@ -1,1 +1,2 @@
 export * from './intMask';
+export * from './intMaskSlug';
