+++ servers/parser-graphql-wrapper/package.json
@@ -69,6 +69,7 @@
     "@types/cors": "^2.8.17",
     "@types/express": "4.17.21",
     "@types/jest": "29.5.12",
+    "@types/md5": "2.3.5",
     "@types/node": "^22.5.2",
     "@types/turndown": "5.0.5",
     "concurrently": "^8.2.2",
